+++ testing/lightcurve_testing.py
@@ -3,4 +3,6 @@ import ATK.Tools as ATK
 SOURCE = 587316166180416640
 
 data = ATK.query("lightcurve", survey="ztf", target=SOURCE)
-# data.show()
+data.show()
+data.plot()
+data.open()